pyproject.toml
[project]
name = "notifyme"
version = "2.2.0"
description = "Windows desktop application to reminds users to blink their eyes every 20 minutes and walk every hour."
readme = "README.md"
requires-python = ">3.13"
dependencies = [
    "pystray>=0.19.5",
    "Pillow>=10.0.0",
    "plyer>=2.1.0",
    "winotify>=1.1.0; sys_platform == 'win32'",
    "pyttsx3>=2.90",
    "setuptools",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["notifyme"]

[dependency-groups]
dev = [
    "pyinstaller>=6.18.0",
    "pre-commit>=3.7.1",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
